requirements.txt
joblib==1.1.0
matplotlib==3.5.2
numpy==1.23.1
opencv_python==4.6.0.66
optuna==2.10.1
Pillow==9.2.0
PyYAML==6.0
scipy==1.8.1
tensorflow==2.9.1
